changes that did nothing.. still cannot send form server side. Perhaps go back to using the send_ack_packet/ send_data_packet and make sure sendto params are correct. Will try tomorrow.

+++ packets.h
@@ -47,7 +47,7 @@ typedef enum{
 		CLOSE,//Close connection request/ack
 }packettype;
 
-const int PTR_SIZE = sizeof(packettype)+sizeof(int)+sizeof(int)+sizeof(PCKLEN);
+const size_t PTR_SIZE = sizeof(packettype)+sizeof(int)+sizeof(int)+sizeof(PCKLEN);
 
 //single packet for handshake requests, acks, and data req's and acks. 
 struct packet{
